% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{canvasBackgroundColor}
\alias{canvasBackgroundColor}
\title{Canvas background colour chartjs plugin}
\usage{
canvasBackgroundColor()
}
\value{
a list
}
\description{
Create an option to colour the canvas background. The function produces a list
that will become an object within the plugins array \code{x$plugins[plugin]}. The id
will become the plugin name, and this fuction takes in a color parameter, thus
\code{x$options$plugins$canvasBackgroundColor$color <- "some colour"}. This is
accessed using \code{cjs_theme(chart.backgroundColor = ...)}
}
