% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs-shiny}
\alias{chartjs-shiny}
\alias{chartjsOutput}
\alias{renderChartjs}
\alias{chartjsProxy}
\title{Shiny bindings for chartjs}
\usage{
chartjsOutput(outputId, width = "100\%", height = "400px")

renderChartjs(expr, env = parent.frame(), quoted = FALSE)

chartjsProxy(
  id,
  session = shiny::getDefaultReactiveDomain(),
  deferUntilFlush = TRUE
)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a chartjs}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id}{element id}

\item{session}{shiny session}

\item{deferUntilFlush}{wait until after the next time Shiny flushes the reactive system}
}
\description{
Output and render functions for using chartjs within Shiny
applications and interactive Rmd documents.
}
