% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\title{Chartjs widget}
\usage{
chartjs(
  data,
  x,
  y,
  type = "bar",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{data.frame with data}

\item{x, y}{column names from data}

\item{type}{chart type}

\item{width, height}{width and height in pixels (optional, defaults to automatic sizing)}

\item{elementId}{id of the widget created with this function}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Basic Chartjs widget
}
