% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_scale_axis_inputs}
\alias{check_scale_axis_inputs}
\title{Check scale_axis inputs}
\usage{
check_scale_axis_inputs(env)
}
\arguments{
\item{env}{function environment, passed on using \code{environment}}
}
\value{
nothing, or an error
}
\description{
Check scale_axis inputs
}
