% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_add_bars.R
\name{cjs_add_bars}
\alias{cjs_add_bars}
\title{Add a dataset to the chartjs bar chart}
\usage{
cjs_add_bars(p, ..., label = NULL, data = NULL)
}
\arguments{
\item{p}{a chartjs htmlwidget}

\item{...}{aesthetic parameters to pass on to chartjs chart types e.g. x, y, TODO, ...}

\item{label}{dataset title (used in legends)}

\item{data}{data from which aesthetic parameters are sourced. If NULL, that within
the chartjs htmlwidget is used.}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Add a dataset to a chartjs instance
}
