% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_add_doughnut_pie.R
\name{cjs_add_doughnut_pie}
\alias{cjs_add_doughnut_pie}
\title{Add a dataset to a chartjs doughnut/pie}
\usage{
cjs_add_doughnut_pie(
  p,
  type = c("doughnut", "pie"),
  ...,
  label = NULL,
  data = NULL
)
}
\arguments{
\item{p}{a chartjs htmlwidget}

\item{type}{which of doughnut', 'pie' wanted}

\item{...}{aesthetic parameters to pass on to chartjs chart types e.g. x, y, colors, TODO, ...}

\item{label}{dataset title (used in legends)}

\item{data}{data from which aesthetic parameters are sourced. If NULL, that within
the chartjs htmlwidget is used.}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Add a dataset to a chartjs instance
}
