% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_add_options.R
\name{cjs_add_options}
\alias{cjs_add_options}
\title{Add/edit options of a chartjs object}
\usage{
cjs_add_options(p, ...)
}
\arguments{
\item{p}{chartjs object}

\item{...}{any chartjs option, using . notation}
}
\value{
An object of class \code{htmlwidget}
}
\description{
A short cut to add options to the chart object. The options object is deeply
nested and has many options. Rather than write a function capturing all the
options (and their children), this function allows you to specify them using
a . (dot) notation e.g. elements.line.tension = 0.25 expands to
\{elements: \{line: \{tension: 0.25\}\}\}. Thus you must determine the options you
desire from the \href{https://www.chartjs.org/docs/latest/configuration/}{chartjs documention}.
Any previously set options should be preserved ...
}
