% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_bar_orientation.R
\name{cjs_bar_orientation}
\alias{cjs_bar_orientation}
\title{Alter the bar orientation}
\usage{
cjs_bar_orientation(p, orientation = c("vertical", "horizontal"))
}
\arguments{
\item{p}{a chartjs object}

\item{orientation}{one of 'vertical' or 'horizontal'}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Alter the bar orientation
}
