% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_cartesian_title.R
\name{cjs_cartesian_title}
\alias{cjs_cartesian_title}
\title{Scale Title Configuration
Namespace: \verb{options.scales[scaleId].title}, it defines options for the scale title. Note that this only applies to cartesian axes.}
\usage{
cjs_cartesian_title(
  display = FALSE,
  align = "center",
  text = "",
  color = htmlwidgets::JS("Chart.defaults.color"),
  font = htmlwidgets::JS("Chart.defaults.font"),
  padding = 4
)
}
\arguments{
\item{display}{boolean FALSE If true, display the axis title.}

\item{align}{string 'center' Alignment of the axis title. Possible options are 'start', 'center' and 'end'}

\item{text}{string|string[] '' The text for the title. (i.e. "# of People" or "Response Choices").}

\item{color}{Color Chart.defaults.color Color of label.}

\item{font}{Font Chart.defaults.font See Fonts}

\item{padding}{Padding 4 Padding to apply around scale labels. Only top, bottom and y are implemented.}
}
\value{
a list
}
\description{
Scale Title Configuration
Namespace: \verb{options.scales[scaleId].title}, it defines options for the scale title. Note that this only applies to cartesian axes.
}
