% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_config_line.R
\name{cjs_config_line}
\alias{cjs_config_line}
\title{Line Configuration}
\usage{
cjs_config_line(
  tension = 0,
  backgroundColor = NULL,
  borderWidth = 3,
  borderColor = NULL,
  borderCapStyle = "butt",
  borderDash = NULL,
  borderDashOffset = NULL,
  borderJoinStyle = "miter",
  capBezierPoints = TRUE,
  cubicInterpolationMode = "default",
  fill = FALSE,
  stepped = FALSE
)
}
\arguments{
\item{tension}{number 0    Bézier curve tension (0 for no Bézier curves).}

\item{backgroundColor}{Color Chart.defaults.backgroundColor    Line fill color.}

\item{borderWidth}{number    3    Line stroke width.}

\item{borderColor}{Color Chart.defaults.borderColor    Line stroke color.}

\item{borderCapStyle}{string    'butt'    Line cap style. See MDN}

\item{borderDash}{number[]    []    Line dash. See MDN}

\item{borderDashOffset}{number    0.0    Line dash offset. See MDN}

\item{borderJoinStyle}{'round'|'bevel'|'miter'    'miter'    Line join style. See MDN}

\item{capBezierPoints}{boolean    true    true to keep Bézier control inside the chart, false for no restriction.}

\item{cubicInterpolationMode}{string    'default'    Interpolation mode to apply. See more...}

\item{fill}{boolean|string    false    How to fill the area under the line. See area charts.}

\item{stepped}{boolean    false    true to show the line as a stepped line (tension will be ignored).}
}
\value{
a list
}
\description{
Line elements are used to represent the line in a line chart.
}
