% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_defaults.R
\name{cjs_defaults}
\alias{cjs_defaults}
\title{Set chartjs defaults}
\usage{
cjs_defaults(...)
}
\arguments{
\item{...}{default options.}
}
\description{
Set global defaults for subsequent chartjs objects in a document. To check
available options, create a chartjs plot, examine in the browser, go to the
javascript console and print the object "Chart.defaults".
}
