% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_events.R
\name{cjs_events}
\alias{cjs_events}
\title{Edit the allowable events on a chartjs object}
\usage{
cjs_events(
  p,
  events = c("mousemove", "mouseout", "click", "touchstart", "touchmove")
)
}
\arguments{
\item{p}{a chartjs htmlwidget}

\item{events}{events to allow}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Allows you to alter/remove events that the chart will respond to
}
\examples{
\dontrun{

// This chart will not respond to mousemove, etc, only clicking
chartjs() \%>\%
  cjs_events('click')

}
}
