% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_example_data.R
\name{cjs_example_data}
\alias{cjs_example_data}
\title{Example data to test and showcase chartjs}
\usage{
cjs_example_data(type = "bar", grouped = FALSE, shuffle = FALSE)
}
\arguments{
\item{type}{data for chart type: 'bar', 'scatter', so far ...}

\item{grouped}{whether to include a group column for multiple datasets/traces}

\item{shuffle}{whether to shuffle the data. Used in testing.}
}
\value{
a data.frame
}
\description{
A function that generates data.frames for input. Scatter XY values use \code{runif(12)}
}
