% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_flexdashboard_example.R
\name{cjs_flexdashboard_example}
\alias{cjs_flexdashboard_example}
\title{Show flexdashboard chartjs example}
\usage{
cjs_flexdashboard_example()
}
\value{
rendered HTML
}
\description{
Show flexdashboard example of chartjs. File path of the Rmd is messaged to the
console.
}
