% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_font.R
\name{cjs_font}
\alias{cjs_font}
\title{Chartjs font options}
\usage{
cjs_font(
  family = "'Helvetica Neue', 'Helvetica', 'Arial', sans-serif",
  size = 12,
  style = "normal",
  weight = NULL,
  lineHeight = 1.2
)
}
\arguments{
\item{family}{string, "'Helvetica Neue', 'Helvetica', 'Arial', sans-serif", Default font
family for all text, follows CSS font-family options.}

\item{size}{number, 12, Default font size (in px) for text. Does not apply to radialLinear scale point labels.}

\item{style}{string, 'normal', Default font style. Does not apply to tooltip title or footer.
Does not apply to chart title. Follows CSS font-style options (i.e. normal, italic, oblique, initial, inherit).}

\item{weight}{string, undefined, Default font weight (boldness). (see MDN).}

\item{lineHeight}{number|string,    1.2, Height of an individual line of text (see MDN).}
}
\value{
a list
}
\description{
There are special global settings that can change all of the fonts on the chart.
These options are in Chart.defaults.font. The global font settings only apply
when more specific options are not included in the config.
}
