% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_grid.R
\name{cjs_grid}
\alias{cjs_grid}
\title{Grid Line Configuration}
\usage{
cjs_grid(
  circular = FALSE,
  color = htmlwidgets::JS("Chart.defaults.borderColor"),
  display = TRUE,
  drawOnChartArea = TRUE,
  drawTicks = TRUE,
  lineWidth = 1,
  offset = FALSE,
  tickBorderDash = NULL,
  tickBorderDashOffset = NULL,
  tickColor = NULL,
  tickLength = 8,
  tickWidth = NULL,
  z = -1
)
}
\arguments{
\item{circular}{boolean FALSE If true, gridlines are circular (on radar and polar area charts only).}

\item{color}{Color Yes Yes}

\item{display}{boolean TRUE If false, do not display grid lines for this axis.}

\item{drawOnChartArea}{boolean TRUE If true, draw lines on the chart area inside the axis lines. This is useful when there are multiple axes and you need to control which grid lines are drawn.}

\item{drawTicks}{boolean TRUE If true, draw lines beside the ticks in the axis area beside the chart.}

\item{lineWidth}{number Yes Yes}

\item{offset}{boolean FALSE If true, grid lines will be shifted to be between labels. This is set to true for a bar chart by default.}

\item{tickBorderDash}{number[] Length and spacing of the tick mark line. If not set, defaults to the grid line borderDash value.}

\item{tickBorderDashOffset}{number Yes Yes}

\item{tickColor}{Color Yes Yes}

\item{tickLength}{number 8 Length in pixels that the grid lines will draw into the axis area.}

\item{tickWidth}{number Yes Yes}

\item{z}{number -1 z-index of the gridline layer. Values <= 0 are drawn under datasets, > 0 on top.}
}
\value{
a list
}
\description{
Namespace: options.scales|scaleId|.grid, it defines options for the grid lines
that run perpendicular to the axis
}
