% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_interaction_mode.R
\name{cjs_interaction_mode}
\alias{cjs_interaction_mode}
\title{Interaction Modes}
\usage{
cjs_interaction_mode(p, mode = c("point", "index", "dataset"))
}
\arguments{
\item{p}{a chartjs htmlwidget}

\item{mode}{type of interaction to allow}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Used for tooltip position mode setting. Should hover show a single data point,
the various groups (if any) at a particular x value, or the dataset of the
datapoint?
}
