% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_list_input_example.R
\name{cjs_list_input_example}
\alias{cjs_list_input_example}
\title{data for list input example}
\usage{
cjs_list_input_example()
}
\value{
a list
}
\description{
list input configuration for passing input 'as-is' to chartjs. Produces a
bar plot with 2 dataset.
}
