% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_scale_axis.R
\name{cjs_scale_axis}
\alias{cjs_scale_axis}
\title{Edit the scale of the plot}
\usage{
cjs_scale_axis(
  p,
  id,
  type = NULL,
  alignToPixels = FALSE,
  backgroundColor = NULL,
  border = cjs_border(),
  display = TRUE,
  grid = cjs_grid(),
  min = NULL,
  max = NULL,
  reverse = FALSE,
  stacked = FALSE,
  suggestedMax = NULL,
  suggestedMin = NULL,
  ticks = cjs_ticks(),
  weight = 0
)
}
\arguments{
\item{p}{plot}

\item{id}{axis id}

\item{type}{string Type of scale being employed. Custom scales can be created and registered with a string key. This allows changing the type of an axis for a chart.}

\item{alignToPixels}{boolean FALSE Align pixel values to device pixels.}

\item{backgroundColor}{Color Background color of the scale area.}

\item{border}{object Border configuration. See \code{cjs_border()}}

\item{display}{boolean|string TRUE Controls the axis global visibility (visible when true, hidden when false). When display: 'auto', the axis is visible only if at least one associated dataset is visible.}

\item{grid}{object Grid line configuration. See \code{cjs_grid()}}

\item{min}{number User defined minimum number for the scale, overrides minimum value from data. more...}

\item{max}{number User defined maximum number for the scale, overrides maximum value from data. more...}

\item{reverse}{boolean FALSE Reverse the scale.}

\item{stacked}{boolean|string FALSE Should the data be stacked. more...}

\item{suggestedMax}{number Adjustment used when calculating the maximum data value. more...}

\item{suggestedMin}{number Adjustment used when calculating the minimum data value. more...}

\item{ticks}{object Tick configuration. See \code{cjs_ticks()}}

\item{weight}{number 0 The weight used to sort the axis. Higher weights are further away from the chart area.}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Common options to all axes. Namespace: options.scales|scaleId| Note: depending
on the axis type used, there will be more options. See \verb{cjs_scale_*} functions.
\code{border}, \code{grid} and \code{ticks} are themselves lists of options
}
