% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_scale_color.R
\name{cjs_scale_color}
\alias{cjs_scale_color}
\title{Scale colour function}
\usage{
cjs_scale_color(
  p,
  backgroundColors,
  borderColors,
  match_background_and_border = FALSE
)
}
\arguments{
\item{p}{chartjs object}

\item{backgroundColors}{vector of colours. Can be named to match the \code{labels}
in the datasets. A single unnamed value will colour all datasets.}

\item{borderColors}{vector of colours. Can be named to match the \code{labels} in the datasets.
A single unnamed value will colour all datasets.}

\item{match_background_and_border}{= FALSE}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Add colours to datasets within a chartjs object
}
