% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_scale_shape.R
\name{cjs_scale_shape}
\alias{cjs_scale_shape}
\title{Scale shape function}
\usage{
cjs_scale_shape(p, pointStyles)
}
\arguments{
\item{p}{chartjs object}

\item{pointStyles}{vector of shapes. See \code{cjs_shapes} for options. Can be named to match the \code{labels}
in the datasets. A single unnamed value will shape all datasets. Uses \code{I()} if
you want points of all datasets a given value to avoid a warning message about
value mismatch to number of datasets}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Shape datasets within a chartjs object.
}
