% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_scale_size.R
\name{cjs_scale_size}
\alias{cjs_scale_size}
\title{Scale size function}
\usage{
cjs_scale_size(p, pointRadii = I(3))
}
\arguments{
\item{p}{chartjs object}

\item{pointRadii}{vector of sizes Can be named to match the \code{labels}
in the datasets. A single unnamed value will size all datasets. Uses \code{I()} if
you want points of all datasets a given value to avoid a warning message about
value mismatch to number of datasets}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Size datasets within a chartjs object.
}
