% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_shapes.R
\docType{data}
\name{cjs_shapes}
\alias{cjs_shapes}
\title{chartjs shape options}
\format{
An object of class \code{list} of length 11.
}
\usage{
cjs_shapes
}
\description{
A named list of chartjs shapes
}
\keyword{datasets}
