% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_shiny_example.R
\name{cjs_shiny_example}
\alias{cjs_shiny_example}
\title{Show Shiny chartjs example}
\usage{
cjs_shiny_example()
}
\value{
rendered HTML
}
\description{
Show Shiny example of chartjs. File path of the Rmd is messaged to the
console.
}
