% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_stack_bars.R
\name{cjs_stack_bars}
\alias{cjs_stack_bars}
\title{Stack datasets to create a stacked barchart}
\usage{
cjs_stack_bars(p)
}
\arguments{
\item{p}{a chartjs object}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Stack datasets to create a stacked barchart. Namespace: \verb{options.scales[scaleId].stacked}
}
