% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_structure_data.R
\name{cjs_structure_data}
\alias{cjs_structure_data}
\title{Structure an R data.frame into a chartjs data structure based on type}
\usage{
cjs_structure_data(data, type)
}
\arguments{
\item{data}{data.frame with columns x, y, TODO ..}

\item{type}{chartjs type}
}
\description{
Different chartjs data structures exist
}
