% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_theme.R
\name{cjs_theme}
\alias{cjs_theme}
\title{Alter chartjs attributes associated with theming}
\usage{
cjs_theme(
  p,
  chart.backgroundColor = NULL,
  chart.borderColor = NULL,
  chart.color = NULL,
  title.text = NULL,
  title.display = FALSE,
  title.align = NULL,
  title.color = NULL,
  title.padding = NULL,
  legend.display = TRUE,
  legend.position = NULL
)
}
\arguments{
\item{p}{a chartjs htmlwidget}

\item{chart.backgroundColor}{background color of the plot}

\item{chart.borderColor}{border color of the plot}

\item{chart.color}{text colour for the plot}

\item{title.text}{plot title}

\item{title.display}{display the plot. Will automatically be TRUE if title text something other than ''}

\item{title.align}{Alignment of the title. Options are: 'start', 'center', 'end'}

\item{title.color}{Color of label}

\item{title.padding}{Padding to apply around scale labels. Only top, bottom and y are implemented}

\item{legend.display}{display the legend}

\item{legend.position}{Position of the legend. Options are: 'top', 'left', 'bottom' 'right'}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Some text
}
