% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_ticks.R
\name{cjs_ticks}
\alias{cjs_ticks}
\title{Common tick options to all axes}
\usage{
cjs_ticks(
  backdropColor = rgb(255, 255, 255, 0.75, maxColorValue = 255),
  backdropPadding = 2,
  callback = NULL,
  display = TRUE,
  color = htmlwidgets::JS("Chart.defaults.color"),
  font = htmlwidgets::JS("Chart.defaults.font"),
  major = NULL,
  padding = 3,
  showLabelBackdrop = FALSE,
  textStrokeColor = NULL,
  textStrokeWidth = 0,
  z = 0
)
}
\arguments{
\item{backdropColor}{Color Yes    'rgba(255, 255, 255, 0.75)'    Color of label backdrops.}

\item{backdropPadding}{Padding 2    Padding of label backdrop.}

\item{callback}{function Returns the string representation of the tick value as it should be displayed on the chart. See callback.}

\item{display}{boolean    true    If true, show tick labels.}

\item{color}{Color Chart.defaults.color    Color of ticks.}

\item{font}{Font    Yes    Chart.defaults.font    See Fonts}

\item{major}{object {}    Major ticks configuration.}

\item{padding}{number 3    Sets the offset of the tick labels from the axis}

\item{showLabelBackdrop}{boolean    Yes    true for radial scale, false otherwise    If true, draw a background behind the tick labels.}

\item{textStrokeColor}{ColorYes``    The color of the stroke around the text.}

\item{textStrokeWidth}{numberYes    0    Stroke width around the text.}

\item{z}{number 0    z-index of tick layer. Useful when ticks are drawn on chart area. Values <= 0 are drawn under datasets, > 0 on top}
}
\value{
a list
}
\description{
These are only the common tick options supported by all axes.
Please see specific axis documentation for all of the available tick options
for that axis. Namespace: options.scales|scaleId|.ticks
}
