% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_update_bars.R
\name{cjs_update_bars}
\alias{cjs_update_bars}
\title{Update chartjs proxy object}
\usage{
cjs_update_bars(proxy, data, x, y)
}
\arguments{
\item{proxy}{\code{chartjsProxy()} object}

\item{data}{data.frame}

\item{x}{x values}

\item{y}{y values}
}
\value{
\code{chartjsProxy()} object
}
\description{
Currently just replaces labels and \verb{datasets[0].data}
}
