% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_labels}
\alias{compare_labels}
\title{Compare chartjs labels}
\usage{
compare_labels(p, labels)
}
\arguments{
\item{p}{chartjs object}

\item{labels}{vector of labels to compare}
}
\description{
Compare chartjs labels
}
