% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_template_literals}
\alias{convert_template_literals}
\title{Convert tooltip template literals}
\usage{
convert_template_literals(string)
}
\arguments{
\item{string}{a string, with or without template literals}
}
\description{
Match R string values to chartjs objects within template literals.
https://www.chartjs.org/docs/latest/configuration/tooltip.html#tooltip-item-context
}
