% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_to_xy_list}
\alias{data_to_xy_list}
\title{Wranlge data to xy list}
\usage{
data_to_xy_list(data)
}
\arguments{
\item{data}{data.frame with x,y columns}
}
\value{
list
}
\description{
Wranlge data to xy list
}
