% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_input_example.R
\name{dots_input_example}
\alias{dots_input_example}
\title{dot input example}
\usage{
dots_input_example()
}
\value{
a list
}
\description{
list input configuration for passing input 'as-is' to chartjs
}
