% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download-js.R
\name{download_chartjs}
\alias{download_chartjs}
\title{Download Chartjs built files}
\usage{
download_chartjs(version = "4.1.1")
}
\arguments{
\item{version}{version to download}
}
\value{
nothing
}
\description{
Download built Chartjs files to package library \code{inst/htmlwidgets/lib/Chart.js-{version}}
}
