% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_chartjs_adapter_date_fns.R
\name{download_chartjs_adapter_date_fns}
\alias{download_chartjs_adapter_date_fns}
\title{Download Chartjs built files}
\usage{
download_chartjs_adapter_date_fns(version = "3.0.0")
}
\arguments{
\item{version}{version to download}
}
\value{
nothing
}
\description{
Download built Chartjs files to package library \code{inst/htmlwidgets/lib/Chart.js-{version}}
}
