% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_dataset_data}
\alias{extract_dataset_data}
\title{Extract elements from chartjs datasets.data}
\usage{
extract_dataset_data(p, element, unique = FALSE)
}
\arguments{
\item{p}{chartjs object}

\item{element}{x or y, or other to select from dtasets.data}

\item{unique}{should the result unlist the element values and unique them?}
}
\value{
a list or vector
}
\description{
Extract elements from chartjs datasets.data
}
