% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_non_null_parameters}
\alias{get_non_null_parameters}
\title{Catpure non-null function parameters}
\usage{
get_non_null_parameters(parameters)
}
\arguments{
\item{parameters}{list captured from \code{as.list(environment())} in function call}
}
\value{
a list
}
\description{
After capturing environment from function, return non-null elements. Also
exclude p and id, plot object and scale id (usually not needed here)
}
