% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_input_example.R
\name{list_input_example}
\alias{list_input_example}
\title{data for list input example}
\usage{
list_input_example()
}
\value{
a list
}
\description{
list input configuration for passing input 'as-is' to chartjs. Produces a
bar plot with 1 dataset.
}
