% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_list}
\alias{modify_list}
\title{Modify list}
\usage{
modify_list(list, new_list)
}
\arguments{
\item{list}{a named list to modify}

\item{new_list}{a new named list to add to \code{list}}
}
\value{
a list
}
\description{
wrapper for ModifyList, in case the initial list is NULL
}
