% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relist_parameter_vector}
\alias{relist_parameter_vector}
\title{Create nested list from vector of value using . in names}
\usage{
relist_parameter_vector(vector)
}
\arguments{
\item{vector}{named vector with '.' delineating nesting}
}
\value{
a list
}
\description{
Create nested list from vector of value using . in names, e.g. a.b -> a$b
}
