% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_new_dataset_ind}
\alias{return_new_dataset_ind}
\title{Return the next dataset index}
\usage{
return_new_dataset_ind(p)
}
\arguments{
\item{p}{chartjs object}
}
\value{
a number
}
\description{
Return the next dataset index
}
