% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_data_from_correct_source}
\alias{select_data_from_correct_source}
\title{Return data.frame given parent function arguments}
\usage{
select_data_from_correct_source(p, data, dots_quo)
}
\arguments{
\item{p}{chartjs object}

\item{data}{data argument passed on from parent function}

\item{dots_quo}{enquos(...) passed on from parent function}
}
\value{
a data.frame
}
\description{
Used within add_* functions. Expect arguments passed on from there. Returns
a data.frame with x, y, ... to be used downstream
}
