% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks_integer_callback.R
\name{ticks_integer_callback}
\alias{ticks_integer_callback}
\title{Callback function for cjs_ticks$callback}
\usage{
ticks_integer_callback(step = 1)
}
\arguments{
\item{jump}{the step value to compare the callback value to determine if a line
shoule be drawn}
}
\value{
a value with a JS_EVAL attribute
}
\description{
The way ticks are produced in chartjs can require a function that determine when
and where they are drawn. This function is a simple callback to produce ticks at
given intervals (step).
}
