% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_scatter_dataset.R
\name{to_scatter_dataset}
\alias{to_scatter_dataset}
\title{Convert xy to chartjs scatter data layout}
\usage{
to_scatter_dataset(data, grouped = FALSE)
}
\arguments{
\item{data}{a data.frame}

\item{grouped}{whether the data has a grouping variable in the column \code{group}}
}
\value{
list
}
\description{
Convert xy to chartjs scatter data layout
}
