% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\title{Chartjs widget}
\usage{
chartjs(..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{...}{list of chartjs options. See \href{https://www.chartjs.org/docs/latest/getting-started/usage.html}{chartjs usage}}

\item{width, height}{width and height in pixels (optional, defaults to automatic sizing)}

\item{elementId}{id of the widget created with this function}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Basic Chartjs widget
}
