% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\title{Chartjs widget}
\usage{
chartjs(
  data = data.frame(),
  ...,
  type = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Can be one of: a data.frame, SharedData or list of chartjs options (for a list,
see \href{https://www.chartjs.org/docs/latest/getting-started/usage.html}{chartjs usage}).}

\item{...}{aesthetic parameters to pass on to chartjs chart types e.g. x, y, colors, TODO, ...}

\item{type}{chartjs type - 'bar', 'scatter', TODO (the others!)}

\item{width, height}{width and height in pixels (optional, defaults to automatic sizing)}

\item{elementId}{id of the widget created with this function}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Basic Chartjs widget
}
